Fixed a case where a single word was being concatenated with itself
There is a special case when comparing equal length strings. As we are using the same
collection of words for the start and end of the concatenation, it is possible that
the same word can match for the prefix and suffix. Therefore a special case is needed
when comparing a single dictionary of words, it ensures the word entry in the dictionary
occurs more than once, otherwise it cannot count towards concatenation.